fix: use non-relative paths in dynamic imports
